moves cli.ts up a level to top level of Src
