+++ src/cli.ts
@@ -1,8 +1,6 @@
 #!/usr/bin/env node
-
-import { clone } from './services/utils.service';
 import * as yargs from 'yargs';
-import { cliOptions, handleArguments } from './cliOptions';
+import { cliOptions, handleArguments } from './cli/cliOptions';
 
 const argv = yargs.options(cliOptions).argv;
 handleArguments(argv);
